requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
openai
pyttsx3
speechrecognition
python-jose
twilio
pymongo
